Separating CRL cache from the persistence layer to ease different implementations persistence methods.
